+++ apps/docs/content/components/accordion/index.ts
@@ -1,0 +1,29 @@
+import usage from "./usage";
+import subtitle from "./subtitle";
+import multiple from "./multiple";
+import compact from "./compact";
+import variants from "./variants";
+import defaultExpandedKeys from "./default-expanded-keys";
+import disabledKeys from "./disabled-keys";
+import startContent from "./start-content";
+import indicator from "./indicator";
+import indicatorFunction from "./indicator-function";
+import customMotion from "./custom-motion";
+import controlled from "./controlled";
+import customStyles from "./custom-styles";
+
+export const accordionContent = {
+  usage,
+  subtitle,
+  multiple,
+  compact,
+  variants,
+  defaultExpandedKeys,
+  disabledKeys,
+  startContent,
+  indicator,
+  indicatorFunction,
+  customMotion,
+  controlled,
+  customStyles,
+};