+++ packages/core/theme/src/components/button.ts
@@ -421,6 +421,29 @@ const button = tv({
 //   xl: "px-8 h-14 text-lg",
 // },
 
+/**
+ * ButtonGroup wrapper **Tailwind Variants** component
+ *
+ * const classNames = buttonGroup({...})
+ *
+ * @example
+ * <div role="group" className={classNames())}>
+ *   // button elements
+ * </div>
+ */
+const buttonGroup = tv({
+  base: "inline-flex items-center justify-center h-auto",
+  variants: {
+    fullWidth: {
+      true: "w-full",
+    },
+  },
+  defaultVariants: {
+    fullWidth: false,
+  },
+});
+
+export type ButtonGroupVariantProps = VariantProps<typeof buttonGroup>;
 export type ButtonVariantProps = VariantProps<typeof button>;
 
-export {button};
+export {button, buttonGroup};