+++ packages/core/theme/src/components/checkbox.ts
@@ -174,7 +174,30 @@ const checkbox = tv({
   },
 });
 
+/**
+ * CheckboxGroup wrapper **Tailwind Variants** component
+ *
+ * const {base, label, wrapper} = checkboxGroup({...})
+ *
+ * @example
+ * <div className={base())}>
+ *  <label className={label()}>Label</label>
+ *  <div className={wrapper()} data-orientation="vertical/horizontal">
+ *     // checkboxes
+ *  </div>
+ * </div>
+ */
+const checkboxGroup = tv({
+  slots: {
+    base: "relative flex flex-col gap-2",
+    label: "relative text-neutral-500",
+    wrapper: "flex flex-col flex-wrap gap-2 data-[orientation=horizontal]:flex-row ",
+  },
+});
+
+export type CheckboxGroupSlots = keyof ReturnType<typeof checkboxGroup>;
+
 export type CheckboxVariantProps = VariantProps<typeof checkbox>;
 export type CheckboxSlots = keyof ReturnType<typeof checkbox>;
 
-export {checkbox};
+export {checkbox, checkboxGroup};