+++ packages/hooks/use-aria-field/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @nextui-org/use-aria-field
 
+## 0.0.0-dev-v2-20230518012432
+
+### Patch Changes
+
+- Some styled were unified in a single file
+- Updated dependencies
+  - @nextui-org/use-aria-slot-id@0.0.0-dev-v2-20230518012432
+  - @nextui-org/use-aria-label@0.0.0-dev-v2-20230518012432
+
+## 0.0.0-dev-v2-20230517170742
+
+### Patch Changes
+
+- Fix build
+- Updated dependencies
+  - @nextui-org/use-aria-slot-id@0.0.0-dev-v2-20230517170742
+  - @nextui-org/use-aria-label@0.0.0-dev-v2-20230517170742
+
 ## 0.0.0-dev-v2-20230515031033
 
 ### Patch Changes